+++ packages/components-native/src/__mocks__/MockModal.tsx
@@ -2,7 +2,9 @@ import React from "react";
 import type { ModalBaseProps } from "react-native";
 import { View } from "react-native";
 
-// @ts-expect-error tsc-ci
-export function MockModal({ visible, children }: ModalBaseProps): JSX.Element {
+export function MockModal({
+  visible,
+  children,
+}: React.PropsWithChildren<ModalBaseProps>) {
   return <View>{visible && children}</View>;
 }