+++ packages/components/src/InputNumber/InputNumber.rebuilt.types.tsx
@@ -1,37 +1,36 @@
-import type { ReactNode } from "react";
 import type React from "react";
-import type { CommonFormFieldProps, FormFieldProps } from "../FormField";
+import type {
+  FocusEvents,
+  HTMLInputBaseProps,
+  KeyboardEvents,
+  RebuiltInputCommonProps,
+} from "../sharedHelpers/types";
 
 export type InputNumberVersion = 1 | 2 | undefined;
 
 export interface InputNumberRebuiltProps
-  extends Omit<CommonFormFieldProps, "version">,
-    Pick<
-      FormFieldProps,
-      "onFocus" | "onBlur" | "inputRef" | "readonly" | "size"
-    > {
-  readonly align?: "center" | "right"; // todo add left and make it default
-  readonly autocomplete?: boolean;
-  readonly autoFocus?: boolean;
+  extends HTMLInputBaseProps,
+    FocusEvents<HTMLInputElement>,
+    KeyboardEvents<HTMLInputElement>,
+    Omit<RebuiltInputCommonProps, "clearable" | "suffix" | "prefix"> {
   readonly defaultValue?: number;
-  readonly description?: ReactNode;
-  readonly error?: string;
   readonly formatOptions?: Intl.NumberFormatOptions;
-  readonly identifier?: string;
-  readonly inline?: boolean;
-  readonly invalid?: boolean;
+
   readonly maxValue?: number;
   readonly minValue?: number;
+
+  /**
+   * Callback for value changes.
+   * @param newValue - The new numeric value
+   * @param event - Optional change event
+   */
   readonly onChange?: (
     newValue: number,
     event?: React.ChangeEvent<HTMLInputElement>,
   ) => void;
+
   readonly showMiniLabel?: boolean;
   readonly value?: number;
-  /**
-   * Version 2 is highly experimental. Avoid using it unless you have talked with Atlantis first.
-   */
-  readonly version: 2;
 }
 
 export interface InputNumberRebuiltRef {