+++ packages/components/src/InputTime/InputTime.types.ts
@@ -1,4 +1,10 @@
 import type { CommonFormFieldProps, FormFieldProps } from "../FormField";
+import type {
+  FocusEvents,
+  HTMLInputBaseProps,
+  KeyboardEvents,
+  RebuiltInputCommonProps,
+} from "../sharedHelpers/types";
 
 export interface InputTimeProps
   extends Pick<
@@ -50,13 +56,25 @@ export interface InputTimeLegacyProps extends InputTimeProps {
 }
 
 export interface InputTimeRebuiltProps
-  extends Omit<
-    InputTimeProps,
-    "defaultValue" | "version" | "validations" | "onValidation"
-  > {
+  extends HTMLInputBaseProps,
+    FocusEvents<HTMLInputElement>,
+    KeyboardEvents<HTMLInputElement>,
+    RebuiltInputCommonProps,
+    Pick<InputTimeProps, "value" | "onChange"> {
   /**
-   * Version 2 is highly experimental, avoid using it unless you have talked with Atlantis first.
+   * Maximum numerical or date value.
    */
-  version: 2;
-  error?: string;
+  readonly max?: number;
+
+  /**
+   * Minimum numerical or date value.
+   */
+  readonly min?: number;
+
+  readonly inputRef?: FormFieldProps["inputRef"];
+
+  /**
+   * @deprecated Use `onKeyDown` or `onKeyUp` instead.
+   */
+  readonly onEnter?: FormFieldProps["onEnter"];
 }