+++ packages/site/src/content/InputDate/exampleV2.ts
@@ -1,0 +1,5 @@
+export default `
+const [date, setDate] = useState(new Date());
+
+return <InputDate version={2} value={date} onChange={setDate} />;
+`;